+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   filterDOMProps.register('customPropToFilter');
   ```
 
+- **Fixed:** Escaping special characters. [\#753](https://github.com/vazco/uniforms/issues/753)
+
 ## [v3.0.0-alpha.5](https://github.com/vazco/uniforms/tree/v3.0.0-alpha.5) (2020-06-17)
 
 - **Breaking:** Removed `modelSync` from `AutoForm` state. [\#739](https://github.com/vazco/uniforms/issues/739)